[fix] 좋아요 api 토글식으로 수정

+++ src/main/java/org/sopt/pawkey/backendapi/domain/post/application/service/PostLikeService.java
@@ -1,12 +1,10 @@
 package org.sopt.pawkey.backendapi.domain.post.application.service;
 
+import org.sopt.pawkey.backendapi.domain.post.api.dto.response.PostLikeResponse;
 import org.sopt.pawkey.backendapi.domain.post.domain.repository.PostLikeRepository;
 import org.sopt.pawkey.backendapi.domain.post.domain.repository.PostRepository;
-import org.sopt.pawkey.backendapi.domain.post.exception.PostLikeBusinessException;
-import org.sopt.pawkey.backendapi.domain.post.exception.PostLikeErrorCode;
 import org.sopt.pawkey.backendapi.domain.post.infra.persistence.entity.PostEntity;
 import org.sopt.pawkey.backendapi.domain.post.infra.persistence.entity.PostLikeEntity;
-import org.sopt.pawkey.backendapi.domain.post.infra.persistence.repository.SpringDataPostRepository;
 import org.sopt.pawkey.backendapi.domain.user.infra.persistence.entity.UserEntity;
 import org.springframework.stereotype.Service;
 
@@ -19,26 +17,22 @@ public class PostLikeService {
 	private final PostLikeRepository postLikeRepository;
 	private final PostRepository postRepository;
 
-	public void like(final UserEntity user, final PostEntity post) {
-		if (postLikeRepository.existsByUserIdAndPostId(user.getUserId(), post.getPostId())) {
-			throw new PostLikeBusinessException(PostLikeErrorCode.DUPLICATE_LIKE);
+	public PostLikeResponse toggleLike(final UserEntity user, final PostEntity post) {
+		boolean isLiked = postLikeRepository.existsByUserIdAndPostId(user.getUserId(), post.getPostId());
+
+		if (isLiked) {
+			postLikeRepository.deleteByUserIdAndPostId(user.getUserId(), post.getPostId());
+			postRepository.decreaseLikeCount(post.getPostId());
+
+			return PostLikeResponse.of(PostLikeResponse.CANCEL_SUCCESS);
+		} else {
+			postLikeRepository.save(PostLikeEntity.builder()
+				.post(post)
+				.user(user)
+				.build());
+			postRepository.increaseLikeCount(post.getPostId());
+
+			return PostLikeResponse.of(PostLikeResponse.LIKE_SUCCESS);
 		}
-
-		postLikeRepository.save(PostLikeEntity.builder()
-			.post(post)
-			.user(user)
-			.build());
-
-		postRepository.increaseLikeCount(post.getPostId());
-	}
-
-	public void cancelLike(UserEntity user, PostEntity post) {
-		int deletedCount = postLikeRepository.deleteByUserIdAndPostId(user.getUserId(), post.getPostId());
-
-		if (deletedCount == 0) {
-			throw new PostLikeBusinessException(PostLikeErrorCode.LIKE_NOT_FOUND);
-		}
-
-		postRepository.decreaseLikeCount(post.getPostId());
 	}
 }
